+++ src/app-api/types/index.ts
@@ -61,6 +61,7 @@ export type {
 } from '../core/tableApi'
 export type { PositionRecord, ViewportApi } from '../core/viewportApi'
 export type {
+  ApplyVisualStyleOptions,
   CreateContinuousMappingOptions,
   VisualPropertyInfo,
   VisualStyleApi,
@@ -149,6 +150,7 @@ export type {
 } from './ElementTypes'
 export {
   ComponentType,
+  MAX_STYLES_PER_NETWORK,
   ValueTypeName,
   VisualPropertyName,
 } from './ElementTypes'